refactor[api-server]: lottery.auth 패키지 위치 변경
