+++ src/main/java/com/watermelon/server/event/lottery/auth/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.lottery.auth.exception;
+package com.watermelon.server.event.lottery.auth.exception;
 
 /**
  * 인증되지 않은 토큰에 대한 예외
