+++ src/main/java/com/watermelon/server/event/lottery/auth/interceptor/LoginCheckInterceptor.java
@@ -1,9 +1,9 @@
-package com.watermelon.server.lottery.auth.interceptor;
+package com.watermelon.server.event.lottery.auth.interceptor;
 
-import com.watermelon.server.lottery.auth.exception.AuthenticationException;
-import com.watermelon.server.lottery.auth.utils.AuthUtils;
-import com.watermelon.server.lottery.auth.exception.InvalidTokenException;
-import com.watermelon.server.lottery.auth.service.TokenVerifier;
+import com.watermelon.server.event.lottery.auth.exception.InvalidTokenException;
+import com.watermelon.server.event.lottery.auth.service.TokenVerifier;
+import com.watermelon.server.event.lottery.auth.utils.AuthUtils;
+import com.watermelon.server.event.lottery.auth.exception.AuthenticationException;
 import com.watermelon.server.lottery.link.service.LinkService;
 import com.watermelon.server.lottery.service.LotteryService;
 import jakarta.servlet.http.HttpServletRequest;
