+++ src/main/java/com/watermelon/server/firebase/FirebaseTokenVerifier.java
@@ -3,8 +3,8 @@ package com.watermelon.server.firebase;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseAuthException;
 import com.google.firebase.auth.FirebaseToken;
-import com.watermelon.server.lottery.auth.exception.InvalidTokenException;
-import com.watermelon.server.lottery.auth.service.TokenVerifier;
+import com.watermelon.server.event.lottery.auth.exception.InvalidTokenException;
+import com.watermelon.server.event.lottery.auth.service.TokenVerifier;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
